refactor endpoint rule node steps
